+++ Lmyc/Controllers/HomeController.cs
@@ -5,9 +5,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Lmyc.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Lmyc.Controllers
 {
+    [Authorize]
     public class HomeController : Controller
     {
         public IActionResult Index()