+++ CHANGELOG.md
@@ -16,6 +16,12 @@ All notable changes to this project will be documented in this file.
 * Use OpenIddict token authentication
 * Add token view models
 
+### 2018-03-22
+### Changed
+* Added DummyData.cs class - RT
+* Direct users to log in when start up the page - RT
+* Changed Boat Des limit to 600 - RT
+
 ## [0.1.2]
 ### 2018-03-20
 #### Changed